+++ lib/cocooned/deprecation.rb
@@ -20,17 +20,21 @@ module Cocooned
   # Deprecated methods
   module Deprecated # :nodoc:
     module Helpers # :nodoc:
-      # @deprecated: Please use {#cocooned_add_item_link} instead
-      def link_to_add_association(*args, &block)
-        cocooned_add_item_link(*args, &block)
-      end
-      deprecate link_to_add_association: 'Use :cocooned_add_link instead', deprecator: Deprecation['3.0']
+      module Tags
+        # @deprecated: Please use {#cocooned_add_item_link} instead
+        def link_to_add_association(*args, &block)
+          cocooned_add_item_link(*args, &block)
+        end
+        deprecate link_to_add_association: 'Use :cocooned_add_link instead',
+                  deprecator: Deprecation['3.0']
 
-      # @deprecated: Please use {#cocooned_remove_item_link} instead
-      def link_to_remove_association(*args, &block)
-        cocooned_remove_item_link(*args, &block)
+        # @deprecated: Please use {#cocooned_remove_item_link} instead
+        def link_to_remove_association(*args, &block)
+          cocooned_remove_item_link(*args, &block)
+        end
+        deprecate link_to_remove_association: 'Use :cocooned_remove_item_link instead',
+                  deprecator: Deprecation['3.0']
       end
-      deprecate link_to_remove_association: 'Use :cocooned_remove_item_link instead', deprecator: Deprecation['3.0']
     end
 
     module TagsHelper # :nodoc:
