+++ lib/cocooned/railtie.rb
@@ -7,7 +7,7 @@ module Cocooned
   class Railtie < ::Rails::Engine # :nodoc:
     initializer 'cocooned.initialize' do |_app|
       ActiveSupport.on_load :action_view do
-        ActionView::Base.include Cocooned::Helpers
+        ActionView::Base.include Cocooned::Helpers::Tags
       end
     end
   end
